python-agents/mercadolivre_context_agent.py
import asyncio
import json
from agents import Agent, Runner, function_tool
from typing import Dict, List, Any
import requests
import os
from dotenv import load_dotenv
load_dotenv()

@function_tool
def get_mercadolivre_categories() -> Dict[str, Any]:
    """Get main product categories available on MercadoLivre"""
    return {
        "categories": [
            "Eletrônicos", "Casa e Jardim", "Moda", "Esportes", 
            "Automotivo", "Livros", "Beleza", "Infantil", 
            "Informática", "Celulares", "Eletrodomésticos"
        ],
        "trending": ["Eletrônicos", "Moda", "Casa e Jardim"],
        "seasonal": "Black Friday approaching - Electronics and Fashion trending"
    }

@function_tool
def get_marketplace_stats() -> Dict[str, Any]:
    """Provide MercadoLivre marketplace statistics and insights"""
    return {
        "active_listings": "18+ million",
        "sellers": "800,000+",
        "countries": ["Brazil", "Argentina", "Mexico", "Colombia", "Chile", "Uruguay"],
        "payment_methods": ["PIX", "Credit Card", "Mercado Pago", "Bank Transfer"],
        "shipping_options": ["Mercado Envios", "Full", "Express", "Standard"],
        "top_selling_categories": ["Electronics", "Home & Garden", "Fashion"],
        "price_range_insights": {
            "electronics": "R$ 50 - R$ 5,000",
            "fashion": "R$ 20 - R$ 500", 
            "home": "R$ 30 - R$ 2,000"
        }
    }

@function_tool
def analyze_user_behavior_patterns() -> Dict[str, Any]:
    """Analyze typical user behavior patterns on MercadoLivre"""
    return {
        "peak_shopping_hours": "19:00 - 22:00",
        "mobile_vs_desktop": "75% mobile, 25% desktop",
        "average_session_duration": "12 minutes",
        "bounce_rate": "32%",
        "conversion_insights": {
            "search_to_purchase": "18%",
            "category_browse_to_purchase": "12%",
            "recommendation_click_rate": "8%"
        },
        "user_preferences": [
            "Free shipping",
            "Fast delivery",
            "Product reviews",
            "Multiple payment options"
        ]
    }

agent = Agent(
    name="MercadoLivre Context Agent",
    instructions="""You are the MercadoLivre Context Agent. Provide comprehensive marketplace data, 
    category information, user behavior insights, and market trends. Help other agents understand 
    the MercadoLivre ecosystem for better decision making and feedback collection.""",
    tools=[get_mercadolivre_categories, get_marketplace_stats, analyze_user_behavior_patterns]
)

async def main():
    result = await Runner.run(agent, "Provide comprehensive MercadoLivre marketplace context")
    print("MercadoLivre Context:", result.final_output)

if __name__ == "__main__":
    asyncio.run(main()) 